Move CredentialsAreNotValidException to storage package to get rid of cyclic dependencies
