+++ server/src/test/java/io/javaclasses/filehub/storage/user/PasswordTest.java
@@ -1,7 +1,6 @@
 package io.javaclasses.filehub.storage.user;
 
 import com.google.common.testing.NullPointerTester;
-import io.javaclasses.filehub.api.user.CredentialsAreNotValidException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
