+++ server/src/test/java/io/javaclasses/filehub/web/serializer/CredentialsAreNotValidExceptionSerializerTest.java
@@ -1,7 +1,7 @@
 package io.javaclasses.filehub.web.serializer;
 
 import com.google.common.testing.NullPointerTester;
-import io.javaclasses.filehub.api.user.CredentialsAreNotValidException;
+import io.javaclasses.filehub.storage.user.CredentialsAreNotValidException;
 import io.javaclasses.filehub.api.user.DataValidationError;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
